@@ -2,18 +2,21 @@ import React from "react";
 export default function AboutMe(props){
     return(
 
-    <div className="content" style={{height:'500px'}}>
-        <div className="col-6">
-        <h1 className="display-2 mx-3">{props.title}</h1>
+    <div>
+        <div className="col-md-6 col-sm">
+        <h1 className="display-2 mx-3 text-center">About Me</h1>
     <p className="text-dark text-left mx-3">
         Hello, I'm Elias Poniridis.
         </p>
         <p className="text-dark text-left mx-3">
-        Licensed Electronics Engineer and researcher based in Athens, Greece, passionated with technology and
-        video games. I have knowledge on game development, Geoinformation and software engineering, 
+            Licensed Electronics Engineer
         </p>
-        <button type="button" className="btn btn-primary">Contact me</button>
-
+        <p className="text-dark text-left mx-3">
+        Based in Athens, Greece, passionated with technology and avid
+        gamer. I have knowledge on game development and software engineering and create Mobile & Web applications and Games. I currently study Acoustics Engineering.
+        </p>
+        <button type="button" className="btn btn-primary mx-2 text-center">Contact me</button>
+            
         </div>
         <div className="col-6">
 
